refactor: avoid repetition and flatten file structure
